Add documentation to API models

+++ api/model/entry.go
@@ -1,13 +1,41 @@
 package model
 
-// Entry contains information about a entry.
+// Entry
+//
+// Contains information about a entry.
+//
+// swagger:model Entry
 type Entry struct {
-	Id            int    `json:"id"`
-	UserId        int    `json:"userId"`
-	StartTime     string `json:"startTime"`
-	EndTime       string `json:"endTime"`
-	BreakDuration int    `json:"breakDuration"`
-	TypeId        int64  `json:"typeId"`
-	ActivityId    int64  `json:"activityId"`
-	Desciption    string `json:"description"`
+	// The ID of the entry.
+	// example: 1
+	Id int `json:"id"`
+
+	// The ID of the user.
+	// example: 1
+	UserId int `json:"userId"`
+
+	// The start time of the entry.
+	// example: 2019-01-01T15:00:00
+	StartTime string `json:"startTime"`
+
+	// The end time of the entry.
+	// example: 2019-01-01T16:00:00
+	EndTime string `json:"endTime"`
+
+	// The break duration of the entry in minutes.
+	// example: 60
+	BreakDuration int `json:"breakDuration"`
+
+	// The ID of the entry type.
+	// example: 1
+	TypeId int `json:"typeId"`
+
+	// The ID of the entry activity.
+	// example: 1
+	ActivityId int `json:"activityId"`
+
+	// The description with additional information about the entry.
+	// min length: 0
+	// max length: 200
+	Desciption string `json:"description"`
 }
